+++ examples/device/dfu_rt/src/main.h
@@ -1,5 +1,0 @@
-#ifndef MAIN_H
-#define MAIN_H
-void led_indicator_pulse(void);
-
-#endif