+++ hw/bsp/mcb1800/mcb1800.c
@@ -37,7 +37,7 @@ void USB0_IRQHandler(void)
   #endif
 
   #if CFG_TUSB_RHPORT0_MODE & OPT_MODE_DEVICE
-    tud_irq_handler(0);
+    tud_int_handler(0);
   #endif
 }
 
@@ -48,7 +48,7 @@ void USB1_IRQHandler(void)
   #endif
 
   #if CFG_TUSB_RHPORT1_MODE & OPT_MODE_DEVICE
-    tud_irq_handler(1);
+    tud_int_handler(1);
   #endif
 }