+++ changelog.md
@@ -4,7 +4,7 @@
 
 ### Breaking
 
-- TinyUSB does not directly implement USB IRQ Handler function anymore. Application must implement IRQ Handler and invoke `tud_irq_handler(rhport)`. This is due to:
+- TinyUSB does not directly implement USB IRQ Handler function anymore. Application must implement IRQ Handler and invoke `tud_int_handler(rhport)`. This is due to:
   
   - IRQ Handler name can be different across system depending on the startup
   - Some OS need to execute enterISR()/exitISR() to work properly, also tracing tool may need to insert trace ISR enter/exit to record usb event
@@ -12,7 +12,7 @@
 
 ### MCU
 
-- All default IRQ Handler is renamed to `dcd_irq_handler()`
+- All default IRQ Handler is renamed to `dcd_int_handler()`
 
 ## 0.6.0 - 2019.03.30